feat: 좋아요 개수 응답 추가
